(Chore) Refactored code
Refactored code to use new authentication to check for invalid request
also the codes were refactored to make them more readable

+++ src/controllers/EmojiController.php
@@ -12,23 +12,22 @@ use Illuminate\Database\Capsule\Manager as DB;
 
 class EmojiController extends BaseController
 {
-    public function __construct()
+    public function __construct($container)
     {
-        parent::__construct();
+        parent::__construct($container);
     }
 
     public function index(ServerRequestInterface $request, ResponseInterface $response)
     {
-        $response = $response->withAddedHeader('Content-type', 'application/json');
+        $response = $response->withAddedHeader('content-type', 'application/json');
         $emojis = Emoji::with('keywords')->get();
+        $body   =    $response->getBody();
 
         $result = array();
-
         foreach ($emojis as $emoji) {
             $result[] = $this->buildEmojiData($emoji);
         }
 
-        $body   =    $response->getBody();
         $body->write(json_encode($result));
         return $response;
     }
@@ -52,46 +51,36 @@ class EmojiController extends BaseController
 
     public function create(ServerRequestInterface $request, ResponseInterface $response)
     {
-            $response   =   $response->withAddedHeader('Content-type', 'application/json');
-            $body       =   $response->getBody();
+        $response   =   $response->withAddedHeader('Content-type', 'application/json');
+        $body       =   $response->getBody();
+        $token      =   $request->getHeader('Authorization')[0];
+        $data       =   $request->getParsedBody();
+        $keywords   =   $data['keywords'];
+        $body       =   $response->getBody();
+        $decodedToken    =   $this->auth->decodeToken($token);
+        $uid        =   $decodedToken->data->uid;
+        $user       =   User::find($uid);
 
-            if($response->getHeader('Expired')[0]  != true) {
-                die();
-                $data       =   $request->getParsedBody();
-                $keywords   =   $data['keywords'];
-                $uid        =   $data['uid'];
-                $body       =   $response->getBody();
-                $message    =   '';
-                $user       =   User::find($uid);
+        $user   =   User::find($uid);
+        if($user == null){
+            throw new \Exception(static::USERDOESNOTEXISTERROR);
+        }
 
-                DB::transaction(function() use($uid, $data, $keywords, $body) {
-                    $user   =   User::find($uid);
+        $message = $this->authenticateRouteRequest($token, $user->jit, $response, static::EMOJICREATED, 201);
 
-                    if($user == null){
-                        throw new \Exception('User does not exist');
-                    }
+        $emoji  =   [
+                'name' => $data['name'],
+                'char' => $data['char'],
+                'category' => $data['category']
+        ];
 
-                    $emoji  =   [
-                        'name' => $data['name'],
-                        'char' => $data['char'],
-                        'category' => $data['category']
-                    ];
-
+        if($message['status'] == 'success') {
+            DB::transaction(function() use($user, $emoji, $keywords) {
                     $emoji = $user->emojis()->create($emoji);
                     foreach ($keywords as $keyword) {
                         $emoji->keywords()->create(['name' => $keyword]);
                     }
-
-                    $message    =   [
-                        'status' => 'Success',
-                        'message' => 'Emoji Created Sucessfully'
-                    ];
-                });
-        } else {
-            $message    =   [
-                        'status' => 'Unauthorized Access',
-                        'message' => 'Token has expired, login to access'
-                    ];
+            });
         }
 
         $body->write(json_encode($message));
@@ -103,38 +92,22 @@ class EmojiController extends BaseController
         $response   =   $response->withAddedHeader('Content-type', 'application/json');
         $body       =   $response->getBody();
         $emoji      =   Emoji::with('keywords')->find($argc['id']);
+        $token      =   $request->getHeader('Authorization')[0];
+        $decodedToken    =   $this->auth->decodeToken($token);
+        $uid        =   $decodedToken->data->uid;
+        $user   =   User::find($uid);
+
+
+        if($user == null){
+            throw new \Exception(static::USERDOESNOTEXISTERROR);
+        }
+
         if ($emoji != null ) {
             $newData        =   $request->getParsedBody();
-
-            DB::transaction(function() use($emoji, $newData) {
-
-                $emoji->name    =   isset($newData['name']) ? $newData['name'] : $emoji->name;
-                $emoji->char    =   isset($newData['char']) ? $newData['char'] : $emoji->char;
-                $emoji->user_id     =   isset($newData['uid'])  ? $newData['uid']  : $emoji->user_id;
-                $emoji->category    =   isset($newData['category']) ? $newData['category'] : $emoji->category;
-
-                foreach ($newData['keywords'] as $key_id => $name) {
-                    $keyword = EmojiKeyword::find($key_id);
-
-                    if(!isset($keyword)) {
-                        $keyword = new EmojiKeyword();
-                        $keyword->emoji_id = $emoji->id;
-                    }
-                    $keyword->name = $name;
-                    $keyword->save();
-                }
-                $emoji->save();
-            });
-            $message        =   [
-                'status' => 'Success',
-                'message' => 'Emoji Updated'
-            ];
-
+            $this->updateEmojiData($emoji, $newData);
+            $message    =   $this->authenticateRouteRequest($token, $user->jit, $response, static::EMOJIUPDATED);
         } else {
-            $message    =   [
-                'status' => 'Error',
-                'message' => 'Emoji not found'
-            ];
+            $message    =   $this->getMessage(static::EMOJINOTFOUNDERROR, $response, 404);
         }
         $body->write(json_encode($message));
         return $response;
@@ -145,23 +118,28 @@ class EmojiController extends BaseController
         $response   =   $response->withAddedHeader('Content-type', 'application/json');
         $body       =   $response->getBody();
         $message    =   [];
-        $emoji     =   Emoji::with('keywords')->find($argc['id']);
+        $emoji      =   Emoji::with('keywords')->find($argc['id']);
+        $token      =   $request->getHeader('Authorization')[0];
+        $decodedToken    =   $this->auth->decodeToken($token);
+        $uid        =   $decodedToken->data->uid;
+        $user   =   User::find($uid);
 
-        if(isset($emoji)) {
-            Db::transaction(function() use ($emoji) {
-                $emoji->delete();
-                $emoji->keywords()->delete();
-            });
 
-            $message = [
-                    "status" => "success",
-                    "message" => "Emoji '".$emoji->name."' deleted!"
-            ];
-        } else {
-            $message = [
-                "status" => "Error",
-                "message" => "Emoji Not found"
-            ];
+        if($user == null){
+            throw new \Exception(static::USERDOESNOTEXISTERROR);
+        }
+
+        $message    =   $this->authenticateRouteRequest($token, $user->jit, $response, "Emoji '".$emoji->name."' deleted!");
+
+        if( $message['status'] == 'success') {
+            if(isset($emoji)) {
+                Db::transaction(function() use ($emoji) {
+                    $emoji->delete();
+                    $emoji->keywords()->delete();
+                });
+            } else {
+                $message    =   $this->getMessage(static::EMOJINOTFOUNDERROR, $response, 404);
+            }
         }
 
         $body->write(json_encode($message));
@@ -188,6 +166,29 @@ class EmojiController extends BaseController
         $result['created_by']       =   $emoji->user->username;
 
         return $result;
+    }
+
+    private function updateEmojiData($emoji, $newData)
+    {
+        DB::transaction(function() use($emoji, $newData) {
+            $emoji->name        =   isset($newData['name']) ? $newData['name'] : $emoji->name;
+            $emoji->char        =   isset($newData['char']) ? $newData['char'] : $emoji->char;
+            $emoji->user_id     =   isset($newData['uid'])  ? $newData['uid']  : $emoji->user_id;
+            $emoji->category    =   isset($newData['category']) ? $newData['category'] : $emoji->category;
+
+            foreach ($newData['keywords'] as $key_id => $name) {
+                $keyword = EmojiKeyword::find($key_id);
+
+                if(!isset($keyword)) {
+                    $keyword = new EmojiKeyword();
+                    $keyword->emoji_id = $emoji->id;
+                }
+                $keyword->name = $name;
+                $keyword->save();
+            }
+            $emoji->save();
+
+        });
     }
